# packages/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/appointments/{id}": {
      "get": {
        "operationId": "AppointmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the appointment to retrieve",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved appointment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Appointment not found or not accessible."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific appointment by ID",
        "tags": [
          "Appointments Management"
        ]
      },
      "put": {
        "operationId": "AppointmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the appointment to update",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update for the appointment. All fields are optional.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              },
              "examples": {
                "Update date and notes": {
                  "summary": "Change date and notes",
                  "value": {
                    "date": "2024-01-10T10:00:00.000Z",
                    "notes": "Updated notes: Patient rescheduled."
                  }
                },
                "Update products": {
                  "summary": "Change associated products",
                  "value": {
                    "products": [
                      {
                        "inventoryItemId": 1,
                        "quantity": 1
                      }
                    ]
                  }
                },
                "Update consultant": {
                  "summary": "Change assigned consultant (Admin only)",
                  "value": {
                    "consultantId": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Appointment not found or not accessible."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing appointment",
        "tags": [
          "Appointments Management"
        ]
      },
      "delete": {
        "operationId": "AppointmentsController_cancelAppointment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the appointment to cancel",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment cancelled successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Appointment not found or not accessible."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an appointment",
        "tags": [
          "Appointments Management"
        ]
      }
    },
...
    "/api/appointment-types": {
      "post": {
        "operationId": "AppointmentTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AppointmentTypes"
        ]
      },
      "get": {
        "operationId": "AppointmentTypesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AppointmentTypes"
        ]
      }
    },
    "/api/appointment-types/{id}": {
      "get": {
        "operationId": "AppointmentTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AppointmentTypes"
        ]
      },
      "put": {
        "operationId": "AppointmentTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AppointmentTypes"
        ]
      },
      "delete": {
        "operationId": "AppointmentTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "AppointmentTypes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Supplement Management API",
    "description": "API documentation for the Supplement Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique user identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "User's full name",
            "example": "Admin User"
          },
          "email": {
            "type": "string",
            "description": "User's unique email address",
            "example": "admin@example.com"
          },
          "role": {
            "type": "string",
            "description": "User's role",
            "enum": [
              "admin",
              "consultant"
            ],
            "example": "admin"
          },
          "status": {
            "type": "string",
            "description": "User's account status",
            "enum": [
              "active",
              "inactive"
            ],
            "example": "active"
          },
          "lastActive": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user's last activity",
            "example": "2025-05-07T13:05:44.879Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was created",
            "example": "2025-05-06T16:20:51.170Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the last update to the user profile",
            "example": "2025-05-07T11:05:44.931Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "status",
          "lastActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "InventoryItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the inventory item",
            "example": 8
          },
          "name": {
            "type": "string",
            "description": "Name of the inventory item",
            "example": "AuditTest Vitamin C (Updated)"
          },
          "description": {
            "type": "string",
            "description": "Description of the item",
            "example": "Vitamin C for audit log testing - updated description",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "description": "Current quantity in stock",
            "example": 90
          },
          "priceWithoutVAT": {
            "type": "number",
            "description": "Price without VAT",
            "example": "10.50",
            "format": "float"
          },
          "vatRate": {
            "type": "number",
            "description": "VAT rate in percentage",
            "example": "21.00",
            "format": "float"
          },
          "priceWithVAT": {
            "type": "number",
            "description": "Calculated price including VAT",
            "example": "12.71",
            "format": "float",
            "nullable": true
          },
          "visibleToAll": {
            "type": "boolean",
            "description": "Indicates if the item is visible to all consultants by default",
            "example": true
          },
          "visibleToSpecificConsultants": {
            "description": "Array of consultant IDs this item is specifically visible to (if visibleToAll is false). Transformed from User[] relation on serialization.",
            "example": [
              13
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the item was created",
            "example": "2025-05-07T08:49:49.782Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the last update to the item",
            "example": "2025-05-07T08:49:56.571Z"
          }
        },
        "required": [
          "id",
          "name",
          "quantity",
          "priceWithoutVAT",
          "vatRate",
          "visibleToAll",
          "visibleToSpecificConsultants",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AppointmentType": {
        "type": "object",
        "properties": {}
      },
      "AppointmentProductItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the appointment product item",
            "example": 1
          },
          "inventoryItemId": {
            "type": "number",
            "description": "ID of the associated inventory item",
            "example": 1
          },
          "inventoryItem": {
            "description": "Details of the associated inventory item",
            "allOf": [
              {
                "$ref": "#/components/schemas/InventoryItem"
              }
            ]
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the item for this appointment",
            "example": 1
          },
          "priceAtTimeOfBooking": {
            "type": "string",
            "description": "Price of one unit at the time of booking",
            "example": "250.00"
          },
          "vatRateAtTimeOfBooking": {
            "type": "string",
            "description": "VAT rate at the time of booking",
            "example": "10.00"
          }
        },
        "required": [
          "id",
          "inventoryItemId",
          "inventoryItem",
          "quantity",
          "priceAtTimeOfBooking",
          "vatRateAtTimeOfBooking"
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the appointment",
            "example": 1
          },
          "patientId": {
            "type": "number",
            "description": "ID of the patient for this appointment",
            "example": 1
          },
          "patient": {
            "description": "Patient details",
            "allOf": [
              {
                "$ref": "#/components/schemas/Patient"
              }
            ]
          },
          "appointmentTypeId": {
            "type": "number",
            "description": "ID of the appointment type",
            "example": 2
          },
          "appointmentType": {
            "description": "Appointment type details",
            "allOf": [
              {
                "$ref": "#/components/schemas/AppointmentType"
              }
            ]
          },
          "consultantId": {
            "type": "number",
            "description": "ID of the consultant for this appointment",
            "example": 2
          },
          "consultant": {
            "description": "Consultant details",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "date": {
            "type": "string",
            "description": "Date and time of the appointment",
            "example": "2025-05-07T21:30:48.000Z",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "description": "Notes for the appointment",
            "example": "Schůzka pro Janu - vitamíny",
            "nullable": true
          },
          "appointmentProducts": {
            "description": "Products associated with the appointment",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppointmentProductItem"
            }
          },
          "totalPrice": {
            "type": "string",
            "description": "Total price of the appointment including products",
            "example": "2775.00",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Status of the appointment",
            "enum": [
              "upcoming",
              "completed",
              "cancelled"
            ],
            "example": "cancelled"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of appointment creation",
            "example": "2025-05-06T17:30:49.008Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of last appointment update",
            "example": "2025-05-06T17:38:50.265Z"
          }
        },
        "required": [
          "id",
          "patientId",
          "patient",
          "appointmentTypeId",
          "appointmentType",
          "consultantId",
          "consultant",
          "date",
          "appointmentProducts",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the patient",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Full name of the patient",
            "example": "Jana Nováková"
          },
          "email": {
            "type": "string",
            "description": "Email address of the patient",
            "example": "jana.novakova@example.com",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the patient",
            "example": "777555333",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Address of the patient",
            "example": "Dlouhá 1, Praha",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Date of birth of the patient",
            "example": "1985-05-15",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about the patient",
            "example": "Aktualizované poznámky k Janě.",
            "nullable": true
          },
          "consultant": {
            "description": "Assigned consultant details",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "consultantId": {
            "type": "number",
            "description": "ID of the assigned consultant",
            "example": 2
          },
          "lastVisit": {
            "type": "string",
            "description": "Date of the patient's last visit",
            "example": "2025-05-08",
            "format": "date",
            "nullable": true
          },
          "totalSpent": {
            "type": "string",
            "description": "Total amount spent by the patient",
            "example": "5825.00"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of patient creation",
            "example": "2025-05-06T16:53:27.179Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of last patient update",
            "example": "2025-05-06T17:30:56.807Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the patient was soft-deleted",
            "nullable": true,
            "example": null
          },
          "purchases": {
            "description": "List of purchases made by the patient",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Purchase"
            }
          },
          "appointments": {
            "description": "List of appointments for the patient",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          }
        },
        "required": [
          "id",
          "name",
          "consultant",
          "consultantId",
          "totalSpent",
          "createdAt",
          "updatedAt"
        ]
      },
      "PurchaseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the purchase item line",
            "example": 2
          },
          "purchaseId": {
            "type": "number",
            "description": "ID of the parent purchase (often redundant if nested)",
            "example": 2
          },
          "inventoryItem": {
            "description": "Details of the purchased inventory item",
            "allOf": [
              {
                "$ref": "#/components/schemas/InventoryItem"
              }
            ]
          },
          "inventoryItemId": {
            "type": "number",
            "description": "ID of the purchased inventory item",
            "example": 1
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the item purchased",
            "example": 2
          },
          "priceAtPurchase": {
            "type": "string",
            "description": "Price per unit at the time of purchase",
            "example": "250.00"
          },
          "vatRateAtPurchase": {
            "type": "string",
            "description": "VAT rate (%) at the time of purchase",
            "example": "10.00"
          },
          "subTotal": {
            "type": "string",
            "description": "Subtotal for this line item (price * quantity with VAT)",
            "example": "550.00"
          }
        },
        "required": [
          "id",
          "inventoryItem",
          "inventoryItemId",
          "quantity",
          "priceAtPurchase",
          "vatRateAtPurchase",
          "subTotal"
        ]
      },
      "Purchase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the purchase",
            "example": 2
          },
          "patient": {
            "description": "Patient who made the purchase",
            "allOf": [
              {
                "$ref": "#/components/schemas/Patient"
              }
            ]
          },
          "patientId": {
            "type": "number",
            "description": "ID of the patient",
            "example": 1
          },
          "consultant": {
            "description": "Consultant who recorded the purchase",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "consultantId": {
            "type": "number",
            "description": "ID of the consultant",
            "example": 2
          },
          "items": {
            "description": "Items included in the purchase",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItem"
            }
          },
          "purchaseDate": {
            "type": "string",
            "description": "Date of the purchase",
            "example": "2025-05-06T19:06:36.140Z",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "string",
            "description": "Total amount of the purchase",
            "example": "550.00"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes for the purchase",
            "example": "Nákup vitamínu C",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of purchase creation",
            "example": "2025-05-06T17:06:36.036Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of last purchase update",
            "example": "2025-05-06T17:06:36.036Z"
          }
        },
        "required": [
          "id",
          "patient",
          "patientId",
          "consultant",
          "consultantId",
          "items",
          "purchaseDate",
          "totalAmount",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateAppointmentDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAppointmentTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the appointment type",
            "example": "Initial Consultation"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the appointment type",
            "example": "First meeting with a new patient."
          },
          "price": {
            "type": "number",
            "description": "Price of the appointment",
            "example": 50,
            "format": "float"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duration of the appointment in minutes",
            "example": 60,
            "minimum": 15
          },
          "visibleToAll": {
            "type": "boolean",
            "description": "Whether this appointment type is visible to all consultants by default",
            "example": true
          },
          "visibleToSpecificConsultantIds": {
            "description": "Array of consultant IDs if appointment type visibility is restricted (used if visibleToAll is false)",
            "example": [
              1,
              5
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "name",
          "price",
          "durationMinutes",
          "visibleToAll"
        ]
      },
      "UpdateAppointmentTypeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}